api_requirements.txt
fastapi>=0.129.0
uvicorn>=0.30.0
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
catboost>=1.2.0
joblib>=1.3.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
